speech/requirements.txt
SpeechRecognition
pyttsx3
requests
python-dotenv
